+++ crates/jarl-core/src/analyze/call.rs
@@ -8,6 +8,7 @@ use crate::lints::any_is_na::any_is_na::any_is_na;
 use crate::lints::browser::browser::browser;
 use crate::lints::download_file::download_file::download_file;
 use crate::lints::duplicated_arguments::duplicated_arguments::duplicated_arguments;
+use crate::lints::expect_null::expect_null::expect_null;
 use crate::lints::expect_true_false::expect_true_false::expect_true_false;
 use crate::lints::grepv::grepv::grepv;
 use crate::lints::length_levels::length_levels::length_levels;
@@ -46,6 +47,9 @@ pub fn call(r_expr: &RCall, checker: &mut Checker) -> anyhow::Result<()> {
         && !checker.should_skip_rule(node, "duplicated_arguments")
     {
         checker.report_diagnostic(duplicated_arguments(r_expr)?);
+    }
+    if checker.is_rule_enabled("expect_null") && !checker.should_skip_rule(node, "expect_null") {
+        checker.report_diagnostic(expect_null(r_expr)?);
     }
     if checker.is_rule_enabled("expect_true_false")
         && !checker.should_skip_rule(node, "expect_true_false")